requirements.txt
streamlit
chromadb
langchain
groq
python-dotenv
sentence-transformers
